net: shell: Add ppp command
Move "ppp" command to a separate file.
